Implement browsing by Tags
- Added new option in Settings
- Browsing by Tags is now possible. Instead of listing directory names, the format is now instead: Artist\Album\Songs

There are still some issues, regarding how MusicBee handles this. Unfortunately, we cannot store the IDs of the retrieved Artists and Albums, and we only get a request as their names from MusicBee. This becomes a problem when you have an Artist with an Album of the same name (e.g. Firewind\Firewind), as it's then not possible to know which one was requested.